pyproject.toml
[project]
name = "seacom-app-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "loguru>=0.7.3",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
    "slowapi>=0.1.9",
    "reportlab>=4.0.0",
    "geoalchemy2>=0.14.0",
    "shapely>=2.0.0",
    "httpx>=0.27.0",
    "strawberry-graphql[fastapi]>=0.235.0",
    "redis>=4.5.0",
]
